refactor: file-logger
