+++ apisix/utils/log-util.lua
@@ -210,7 +210,27 @@ function _M.inject_get_full_log(f)
 end
 
 
+local function is_match(matches, ctx)
+    local match_result
+    for _, match in pairs(matches) do
+        local expr, _ = expr.new(match)
+        match_result = expr:eval(ctx.var)
+        if match_result then
+            break
+        end
+    end
+
+    return match_result
+end
+
+
 function _M.get_log_entry(plugin_name, conf, ctx)
+    -- If the "matches" configuration is set and the matching conditions are not met,
+    -- then do not log the message.
+    if conf.matches and not is_match(conf.matches, ctx) then
+        return
+    end
+
     local metadata = plugin.plugin_metadata(plugin_name)
     core.log.info("metadata: ", core.json.delay_encode(metadata))
 
@@ -265,6 +285,12 @@ function _M.check_log_schema(conf)
         local ok, err = expr.new(conf.include_resp_body_expr)
         if not ok then
             return nil, "failed to validate the 'include_resp_body_expr' expression: " .. err
+        end
+    end
+    if conf.matches then
+        local ok, err = expr.new(conf.matches)
+        if not ok then
+            return nil, "failed to validate the 'matches' expression: " .. err
         end
     end
     return true, nil
